spark/requirements.txt
pyspark==3.1.2
delta-spark==1.0.0